Move NumberBox and add new constructor
